Android 13 Background Support, Upgrade Gradle.
Implemented permission support for BODY_SENSORS_BACKGROUND, allowing for the Sensor Data permission "All the time". This prevents data from not being sent when the screen turns off on certain WearOS 4 devices.

Changed target API to 33 (Tiramisu), (Minimum SDK is still 30).

Updated gradle for API 33, unsure if necessary.

Added a probably unneeded SuppressLint to updateHeartRate, as Android Studio counted it as an error.